+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/schema/TimeSeriesCountOperator.java
@@ -87,7 +87,7 @@ public class TimeSeriesCountOperator implements SourceOperator {
   public TsBlock next() {
     isFinished = true;
     TsBlockBuilder tsBlockBuilder = new TsBlockBuilder(outputDataTypes);
-    int count = 0;
+    long count = 0;
     try {
       if (key != null && value != null) {
         count =
@@ -104,7 +104,7 @@ public class TimeSeriesCountOperator implements SourceOperator {
       throw new RuntimeException(e.getMessage(), e);
     }
     tsBlockBuilder.getTimeColumnBuilder().writeLong(0L);
-    tsBlockBuilder.getColumnBuilder(0).writeInt(count);
+    tsBlockBuilder.getColumnBuilder(0).writeLong(count);
     tsBlockBuilder.declarePosition();
     return tsBlockBuilder.build();
   }
